combine common_web and common_tweb now that there is no nevow renderer
